backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
youtube-transcript-api==1.2.4
yt-dlp>=2025.1.1
pydantic>=2.11.0
pydantic-settings>=2.7.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis>=5.2.0
celery>=5.4.0
httpx>=0.27.2
python-multipart>=0.0.12
groq>=0.13.1
anthropic>=0.39.0
python-dotenv==1.0.1
aiofiles>=24.1.0
python-docx>=1.1.0
